+++ src/main/java/com/NewEmployeeManagement/Service/EmployeeService.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface EmployeeService {
-    Employee createEmployee(EmployeeCreateDTO dto, String role, String email, int departmentId, Long categoryId,
+    Employee createEmployee(EmployeeCreateDTO dto, String role, String email, Long departmentId, Long categoryId,
                             MultipartFile idProof,
                             MultipartFile employeePhoto,
                             MultipartFile resume,
@@ -18,18 +18,18 @@ public interface EmployeeService {
 
     List<Employee> getAllEmployees(String role, String email);
 
-    EmployeeResponseDTO getEmployeeById(int id, String role, String email);
+    EmployeeResponseDTO getEmployeeById(Long id, String role, String email);
 
     Employee updateEmployee(
-            int id, EmployeeCreateDTO dto, String role,
-            String email, int departmentId,Long categoryId, MultipartFile idProof,
+            Long id, EmployeeCreateDTO dto, String role,
+            String email, Long departmentId,Long categoryId, MultipartFile idProof,
             MultipartFile employeePhoto, MultipartFile resume, MultipartFile addressProof, MultipartFile experienceLetter
     );
 
-    void deleteEmployee(int id, String role, String email);
+    void deleteEmployee(Long id, String role, String email);
     void carryForwardLeavesForEligibleEmployees();
 
     Map<String, Object> getCrudPermissionForEmployeeByEmail(String empEmail);
 
-    Employee updateStatus(Integer id, String status);
+    Employee updateStatus(Long id, String status);
 }
